Refactored files for electron builder to work
