+++ app/nativeAPI/ForegroundWindow.ts
