+++ app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "auto-hot-pie-angular",
+  "name": "autohotpie",
   "version": "3.0.0",
   "description": "",
   "homepage": "https://github.com/dumbeau/AutoHotPie",
