+++ src/app/home/pie-menu-list/pie-menu-list-row/pie-menu-list-row.component.ts
@@ -1,6 +1,6 @@
 import {Component, EventEmitter, Input, Output, ViewChild} from '@angular/core';
 import {NbPosition} from '@nebular/theme';
-import {PieMenu} from '../../../../preferences/PieMenu';
+import {PieMenu} from '../../../../../app/preferences/PieMenu';
 
 @Component({
   selector: 'app-pie-menu-list-row',
